+++ ichnaea/api/locate/views.py
@@ -21,13 +21,14 @@ class BaseLocateView(BaseAPIView):
 
         query = Query(
             fallback=request_data.get('fallbacks'),
-            geoip=self.request.client_addr,
+            ip=self.request.client_addr,
             cell=request_data.get('cell'),
             wifi=request_data.get('wifi'),
             api_key=api_key,
             api_name=self.view_name,
             api_type=self.view_type,
             session=self.request.db_ro_session,
+            geoip_db=self.request.registry.geoip_db,
             stats_client=self.stats_client,
         )